make serialize_headers & deserialize_headers public
