+++ ext/cache/sqlite.rs
@@ -21,7 +21,9 @@ use std::sync::Arc;
 use std::time::SystemTime;
 use std::time::UNIX_EPOCH;
 
+use crate::deserialize_headers;
 use crate::get_header;
+use crate::serialize_headers;
 use crate::vary_header_matches;
 use crate::Cache;
 use crate::CacheDeleteRequest;
@@ -419,42 +421,4 @@ impl Resource for CacheResponseResource {
 pub fn hash(token: &str) -> String {
   use sha2::Digest;
   format!("{:x}", sha2::Sha256::digest(token.as_bytes()))
-}
-
-fn serialize_headers(headers: &[(ByteString, ByteString)]) -> Vec<u8> {
-  let mut serialized_headers = Vec::new();
-  for (name, value) in headers {
-    serialized_headers.extend_from_slice(name);
-    serialized_headers.extend_from_slice(b"\r\n");
-    serialized_headers.extend_from_slice(value);
-    serialized_headers.extend_from_slice(b"\r\n");
-  }
-  serialized_headers
-}
-
-fn deserialize_headers(
-  serialized_headers: &[u8],
-) -> Vec<(ByteString, ByteString)> {
-  let mut headers = Vec::new();
-  let mut piece = None;
-  let mut start = 0;
-  for (i, byte) in serialized_headers.iter().enumerate() {
-    if byte == &b'\r' && serialized_headers.get(i + 1) == Some(&b'\n') {
-      if piece.is_none() {
-        piece = Some(start..i);
-      } else {
-        let name = piece.unwrap();
-        let value = start..i;
-        headers.push((
-          ByteString::from(&serialized_headers[name]),
-          ByteString::from(&serialized_headers[value]),
-        ));
-        piece = None;
-      }
-      start = i + 2;
-    }
-  }
-  assert!(piece.is_none());
-  assert_eq!(start, serialized_headers.len());
-  headers
 }
